requirements.txt
streamlit
plotly
pandas
sqlalchemy
pyodbc
python-dotenv